Move to_unicode to generic utils
